MAIN: More refactoring and import statement changes
