+++ zipline/finance/blotter/simulated_blotter.py
@@ -19,7 +19,7 @@ from copy import copy
 from six import iteritems
 
 from zipline.assets import Equity, Future, Asset
-from zipline.finance.blotter.blotter import Blotter
+from zipline.finance.blotter import Blotter, register
 from zipline.finance.order import Order
 from zipline.finance.slippage import (
     DEFAULT_FUTURE_VOLUME_SLIPPAGE_BAR_LIMIT,
@@ -32,14 +32,13 @@ from zipline.finance.commission import (
     PerContract,
     PerShare,
 )
-from zipline.finance.blotter.core import register_blotter_class
 from zipline.utils.input_validation import expect_types
 
 log = Logger('Blotter')
 warning_logger = Logger('AlgoWarning')
 
 
-@register_blotter_class('default')
+@register('default')
 class SimulatedBlotter(Blotter):
     def __init__(self, data_frequency, equity_slippage=None,
                  future_slippage=None, equity_commission=None,
